# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: LazyCloud API
  version: "1"
servers:
  - url: /
security:
  - bearer: []
  - session: []
paths:
...
  /v1/workspaces/{workspace}/apps/{app}/previews:
    parameters:
      - $ref: "#/components/parameters/WorkspacePath"
      - $ref: "#/components/parameters/AppPath"
    post:
      operationId: createPreview
      summary: Start a preview container for one workload definition
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PreviewRequest" }
      responses:
        "201":
          description: The preview; its container starts asynchronously
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Preview" }
        default: { $ref: "#/components/responses/Error" }
  /v1/workspaces/{workspace}/previews/{preview}:
    parameters:
      - $ref: "#/components/parameters/WorkspacePath"
      - $ref: "#/components/parameters/PreviewPath"
    get:
      operationId: getPreview
      summary: Read a preview, optionally waiting until its container is ready
      parameters:
        - name: wait_seconds
          in: query
          description: Hold the request until the container is ready or the preview stops, or this many seconds pass.
          schema: { type: integer, minimum: 0, maximum: 60, default: 0 }
      responses:
        "200":
          description: The preview
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Preview" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: stopPreview
      summary: Stop the preview and its container
      responses:
        "200":
          description: The stopped preview
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Preview" }
        default: { $ref: "#/components/responses/Error" }
  /v1/workspaces/{workspace}/previews/{preview}/files:
    parameters:
      - $ref: "#/components/parameters/WorkspacePath"
      - $ref: "#/components/parameters/PreviewPath"
    post:
      operationId: syncPreviewFiles
      summary: Write and remove files in the preview container's workspace
      description: >-
        The body is a tar archive of the files to write, paths relative to
        the workspace, at most 64 MiB. A path to remove is a regular entry of size zero carrying
        the PAX record `LAZYCLOUD.removed=1`. The container's runners restart
        once their current work finishes.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        "200":
          description: What was applied
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PreviewSync" }
        default: { $ref: "#/components/responses/Error" }
  /v1/workspaces/{workspace}/previews/{preview}/output:
    parameters:
      - $ref: "#/components/parameters/WorkspacePath"
      - $ref: "#/components/parameters/PreviewPath"
    get:
      operationId: streamPreviewOutput
      summary: The preview container's output after a cursor, one JSON ContainerLogEntry per line
      description: >-
        A followed stream keeps the preview alive, writes a blank line after
        15 seconds without entries and ends when the preview stops. Without a
        follower the preview stops a minute later.
      parameters:
        - name: after
          in: query
          schema: { type: integer, format: int64, minimum: 0, default: 0 }
        - name: follow
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Newline-delimited ContainerLogEntry objects
          content:
            application/x-ndjson:
              schema: { type: string, format: binary }
        default: { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
    session:
      description: >-
        The dashboard's browser session, set by GitHub sign-in. Requests
        that change state with it must send an Origin header naming the
        dashboard.
      type: apiKey
      in: cookie
      name: __Host-lazycloud_session
  parameters:
    WorkspacePath:
      name: workspace
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Name" }
    AppPath:
      name: app
      in: path
      required: true
      schema: { $ref: "#/components/schemas/AppName" }
...
    PreviewPath:
      name: preview
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: A typed error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message: { type: string }
    ErrorCode:
      type: string
      enum:
        - invalid_request
        - unauthenticated
        - forbidden
        - not_found
        - conflict
        - payload_too_large
        - unsupported
        - too_many_pending_tasks
        - task_not_finished
        - unavailable
        - internal
    Name:
      type: string
      pattern: "^[a-z][a-z0-9-]{0,62}$"
    AppName:
      type: string
      pattern: "^[a-z][a-z0-9_]{0,62}$"
    WorkloadName:
      type: string
      pattern: "^[A-Za-z_][A-Za-z0-9_-]{0,62}$"
    Sha256:
      description: Lowercase hex SHA-256 digest
      type: string
      pattern: "^[0-9a-f]{64}$"
...
    FunctionSpec:
      type: object
      required: [name, handler, source, image, resources]
      properties:
        name: { $ref: "#/components/schemas/WorkloadName" }
        handler:
          description: "module:qualname within the source archive"
          type: string
          pattern: "^[A-Za-z_][A-Za-z0-9_.]*:[A-Za-z_][A-Za-z0-9_.]*$"
        source:
          $ref: "#/components/schemas/SourceRef"
        image:
          $ref: "#/components/schemas/ImageSpec"
        resources:
          $ref: "#/components/schemas/Resources"
        timeout_seconds:
          type: integer
          minimum: 1
          maximum: 86400
          default: 3600
        retry_policy:
          $ref: "#/components/schemas/RetryPolicy"
        concurrency:
          description: Tasks one container runs at once, one process per slot.
          type: integer
          minimum: 1
          maximum: 256
          default: 1
        # Defaults: 10, 0 for a scheduled function, and -1 whenever
        # autoscaler.min_containers is above zero. -1 leaves idle containers
        # above the minimum to the planner, which stops them at once.
        keep_warm_seconds:
          description: Idle seconds before a container above the minimum stops.
          type: integer
          minimum: -1
          maximum: 86400
        autoscaler:
          $ref: "#/components/schemas/Autoscaler"
        max_pending_tasks:
          type: integer
          minimum: 1
          maximum: 1000000
          default: 100
        environment:
          type: object
          additionalProperties: { type: string }
        # Five cron fields, an alias such as @hourly, or `every N` minutes
        # (m, at most 59), hours (h, at most 23) or days (d, at most 31).
        # Deploy stores the normalized expression.
        cron:
          description: Run the function on this UTC schedule.
          type: string
          minLength: 1
          maxLength: 160
        secrets:
          description: Secrets the container receives as environment variables of the same name.
          type: array
          maxItems: 100
          uniqueItems: true
          items: { $ref: "#/components/schemas/SecretName" }
        # http or https, with a host and without credentials or a fragment.
        callback_url:
          description: Receives a signed POST when a task is retried or finishes.
          type: string
          maxLength: 2048
        in_process:
          description: Run the concurrency slots as threads of one runner process.
          type: boolean
          default: false
        lifecycle_hooks:
          $ref: "#/components/schemas/LifecycleHooks"
        client_contract:
          description: The signature `lazycloud app export` types clients from.
          type: object
          additionalProperties: true
          x-go-type: json.RawMessage
        http:
          $ref: "#/components/schemas/HttpSpec"
    SourceRef:
      type: object
      required: [sha256]
      properties:
        sha256: { $ref: "#/components/schemas/Sha256" }
    ImageSpec:
      type: object
      required: [python_version]
      properties:
        python_version:
          description: The Python minor version the runtime is mounted for.
          type: string
          enum: ["3.10", "3.11", "3.12", "3.13", "3.14"]
        # A ready image the workspace resolved, built for python_version.
        # Without one the function runs the platform's image for it.
        image_id: { $ref: "#/components/schemas/ImageId" }
        reference:
          description: The image by digest the release runs, set by the server.
          type: string
          readOnly: true
    Resources:
      description: >-
        Reservations the container always keeps. CPU above the reservation is
        shared up to `cpu_limit_millis`, by default the reservation plus 16
        cores. Memory above the reservation is allowed up to
        `memory_limit_mib`, by default four times the reservation, at least
        1 GiB and at most 8 GiB above it; the container is killed beyond it.
      type: object
      required: [cpu_millis, memory_mib]
      properties:
        cpu_millis:
          type: integer
          minimum: 125
          maximum: 192000
        cpu_limit_millis:
          type: integer
          minimum: 125
          maximum: 192000
        memory_mib:
          type: integer
          minimum: 128
          maximum: 1572864
        memory_limit_mib:
          type: integer
          minimum: 128
          maximum: 1572864
    RetryPolicy:
      type: object
      required: [max_attempts]
      properties:
        max_attempts:
          description: Attempts including the first.
          type: integer
          minimum: 1
          maximum: 100
        delay_seconds:
          type: number
          minimum: 0
          maximum: 3600
          default: 0
        backoff:
          type: string
          enum: [fixed, exponential]
          default: fixed
        max_delay_seconds:
          type: number
          minimum: 0
          maximum: 86400
    Autoscaler:
      type: object
      properties:
        min_containers:
          type: integer
          minimum: 0
          maximum: 1000
          default: 0
        max_containers:
          type: integer
          minimum: 1
          maximum: 1000
          default: 1
        tasks_per_container:
          type: integer
          minimum: 1
          maximum: 10000
          default: 1
...
    LifecycleHooks:
      description: >-
        Callables the runner invokes with a context object, each a
        `module:qualname` reference into the source. on_start runs once per
        runner process after the handler loads, and its failure is a load
        error. The others run in the container around each attempt, in
        order; their failures are logged and do not change the outcome.
      type: object
      properties:
        on_start: { $ref: "#/components/schemas/HookReferences" }
        on_running: { $ref: "#/components/schemas/HookReferences" }
        on_success: { $ref: "#/components/schemas/HookReferences" }
        on_error: { $ref: "#/components/schemas/HookReferences" }
        on_retry: { $ref: "#/components/schemas/HookReferences" }
        on_failure: { $ref: "#/components/schemas/HookReferences" }
        on_finish: { $ref: "#/components/schemas/HookReferences" }
    HookReferences:
      type: array
      maxItems: 16
      items:
        type: string
        pattern: "^[A-Za-z_][A-Za-z0-9_.]*:[A-Za-z_][A-Za-z0-9_.]*$"
    SecretName:
      description: An environment variable name; the LAZYCLOUD_ prefix is reserved.
      type: string
      pattern: "^[A-Za-z_][A-Za-z0-9_]{0,239}$"
...
    ImageId:
      type: string
      pattern: "^img_[0-9a-f]{24}$"
...
    HttpSpec:
      description: >-
        Makes the workload answer HTTP. Its slots are `workers` processes,
        each handling `concurrency` requests at once.
      type: object
      required: [kind]
      properties:
        kind: { $ref: "#/components/schemas/HttpKind" }
        route:
          description: The path an endpoint answers on.
          type: string
          pattern: "^/[^\\s?#]*$"
          maxLength: 512
          default: /
        methods:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/HttpMethod" }
          default: [GET, POST]
        domain:
          $ref: "#/components/schemas/Hostname"
        authorized:
          description: Require a bearer token for the workspace.
          type: boolean
          default: true
        workers:
          type: integer
          minimum: 1
          maximum: 64
          default: 1
    HttpKind:
      type: string
      enum: [endpoint, asgi, realtime]
    HttpMethod:
      type: string
      enum: [DELETE, GET, HEAD, OPTIONS, PATCH, POST, PUT, TRACE]
    Hostname:
      type: string
      maxLength: 253
      pattern: "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?(?:\\.[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?)+$"
...
    PreviewRequest:
      type: object
      required: [spec]
      properties:
        spec: { $ref: "#/components/schemas/FunctionSpec" }
        timeout_seconds:
          description: Stop after this long; 0 runs until stopped.
          type: integer
          minimum: 0
          maximum: 86400
          default: 0
    PreviewState:
      type: string
      enum: [starting, ready, stopped]
    PreviewKind:
      type: string
      enum: [function, endpoint, asgi, realtime]
    Preview:
      type: object
      required: [id, app, name, kind, state, url, created_at]
      properties:
        id: { type: string, format: uuid }
        app: { $ref: "#/components/schemas/AppName" }
        name: { $ref: "#/components/schemas/WorkloadName" }
        kind: { $ref: "#/components/schemas/PreviewKind" }
        state: { $ref: "#/components/schemas/PreviewState" }
        container_id: { type: string, format: uuid }
        url:
          description: Where the preview answers; a function preview takes tasks through the API.
          type: string
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    PreviewSync:
      type: object
      required: [written, removed]
      properties:
        written: { type: integer }
        removed: { type: integer }
...